+++ c/cov.c
@@ -35,7 +35,7 @@ int cov_s (float *Y, const float *X1, const float *X2, const size_t R1, const si
     const size_t L1 = (dim==0u) ? R1 : (dim==1u) ? C1 : (dim==2u) ? S1 : H1;
     const size_t L2 = (dim==0u) ? R2 : (dim==1u) ? C2 : (dim==2u) ? S2 : H2;
     if (L1!=L2) { fprintf(stderr,"error in cov_s: vectors in X1 and X2 must have the same length\n"); return 1; }
-    const float den = 1.0f/L, den2 = (biased) ? den : 1.0f/(L-1u);
+    const float den = 1.0f/(float)L, den2 = (biased) ? den : 1.0f/(float)(L-1u);
     float mn1 = 0.0f, mn2 = 0.0f, sm2 = 0.0f;
 
     if (N==0u) {}
@@ -107,7 +107,7 @@ int cov_d (double *Y, const double *X1, const double *X2, const size_t R1, const
     const size_t L1 = (dim==0u) ? R1 : (dim==1u) ? C1 : (dim==2u) ? S1 : H1;
     const size_t L2 = (dim==0u) ? R2 : (dim==1u) ? C2 : (dim==2u) ? S2 : H2;
     if (L1!=L2) { fprintf(stderr,"error in cov_s: vectors in X1 and X2 must have the same length\n"); return 1; }
-    const double den = 1.0/L, den2 = (biased) ? den : 1.0/(L-1u);
+    const double den = 1.0/(double)L, den2 = (biased) ? den : 1.0/(double)(L-1u);
     double mn1 = 0.0, mn2 = 0.0, sm2 = 0.0;
 
     if (N==0u) {}
@@ -179,7 +179,7 @@ int cov_c (float *Y, const float *X1, const float *X2, const size_t R1, const si
     const size_t L1 = (dim==0u) ? R1 : (dim==1u) ? C1 : (dim==2u) ? S1 : H1;
     const size_t L2 = (dim==0u) ? R2 : (dim==1u) ? C2 : (dim==2u) ? S2 : H2;
     if (L1!=L2) { fprintf(stderr,"error in cov_c: vectors in X1 and X2 must have the same length\n"); return 1; }
-    const float den = 1.0f/L, den2 = (biased) ? den : 1.0f/(L-1u);
+    const float den = 1.0f/(float)L, den2 = (biased) ? den : 1.0f/(float)(L-1u);
     float mn1r, mn1i, mn2r, mn2i, x1r, x1i, x2r, x2i, yr, yi;
 
     if (N==0u) {}
@@ -285,7 +285,7 @@ int cov_z (double *Y, const double *X1, const double *X2, const size_t R1, const
     const size_t L1 = (dim==0u) ? R1 : (dim==1u) ? C1 : (dim==2u) ? S1 : H1;
     const size_t L2 = (dim==0u) ? R2 : (dim==1u) ? C2 : (dim==2u) ? S2 : H2;
     if (L1!=L2) { fprintf(stderr,"error in cov_z: vectors in X1 and X2 must have the same length\n"); return 1; }
-    const double den = 1.0/L, den2 = (biased) ? den : 1.0/(L-1u);
+    const double den = 1.0/(double)L, den2 = (biased) ? den : 1.0/(double)(L-1u);
     double mn1r, mn1i, mn2r, mn2i, x1r, x1i, x2r, x2i, yr, yi;
 
     if (N==0u) {}
