+++ c/gscore.c
@@ -21,7 +21,7 @@ int gscore_s (float *X, const size_t R, const size_t C, const size_t S, const si
     const size_t N = R*C*S*H;
     const size_t L = (dim==0u) ? R : (dim==1u) ? C : (dim==2u) ? S : H;
     if (L<2u) { fprintf(stderr,"error in gscore_s: L (vec length) must be > 1\n"); return 1; }
-    const float den = 1.0f/L, den2 = den;
+    const float den = 1.0f/(float)L, den2 = den;
     float mn = 0.0f, sd = 0.0f;
 
     if (N==0u) {}
@@ -79,7 +79,7 @@ int gscore_d (double *X, const size_t R, const size_t C, const size_t S, const s
     const size_t N = R*C*S*H;
     const size_t L = (dim==0u) ? R : (dim==1u) ? C : (dim==2u) ? S : H;
     if (L<2u) { fprintf(stderr,"error in gscore_s: L (vec length) must be > 1\n"); return 1; }
-    const double den = 1.0/L, den2 = den;
+    const double den = 1.0/(double)L, den2 = den;
     double mn = 0.0, sd = 0.0;
 
     if (N==0u) {}
